requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.2
click==8.1.7
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
motor==3.3.1
nest-asyncio==1.6.0
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.8.0
tzdata==2024.1
uvicorn==0.23.2
wcwidth==0.2.13
zipp==3.17.0
